Change amount type from Int to Double to handle decimal digits

+++ DigibankLight/DigibankLight/Modules/Dashboard/Model/Transactions/TransactionsResponse.swift
@@ -9,7 +9,7 @@ struct TransactionsResponse: Equatable {
     
     struct Transaction: Equatable {
         let transactionId: String?
-        let amount: Int?
+        let amount: Double?
         let transactionDate: Date?
         let description: String?
         let transactionType: String?
